move EngineType and MetricType to MetaTypes.h
