+++ core/src/db/engine/ExecutionEngineImpl.h
@@ -78,10 +78,11 @@ class ExecutionEngineImpl : public ExecutionEngine {
 
     Status
     ExecBinaryQuery(query::GeneralQueryPtr general_query, faiss::ConcurrentBitsetPtr& bitset,
-                    std::unordered_map<std::string, DataType>& attr_type, std::string& vector_placeholder) override;
+                    std::unordered_map<std::string, meta::hybrid::DataType>& attr_type,
+                    std::string& vector_placeholder) override;
 
     Status
-    HybridSearch(scheduler::SearchJobPtr job, std::unordered_map<std::string, DataType>& attr_type,
+    HybridSearch(scheduler::SearchJobPtr job, std::unordered_map<std::string, meta::hybrid::DataType>& attr_type,
                  std::vector<float>& distances, std::vector<int64_t>& search_ids, bool hybrid) override;
 
     Status
@@ -128,10 +129,10 @@ class ExecutionEngineImpl : public ExecutionEngine {
 
     Status
     ProcessTermQuery(faiss::ConcurrentBitsetPtr& bitset, query::GeneralQueryPtr general_query,
-                     std::unordered_map<std::string, DataType>& attr_type);
+                     std::unordered_map<std::string, meta::hybrid::DataType>& attr_type);
 
     Status
-    ProcessRangeQuery(const engine::DataType data_type, const std::string& operand,
+    ProcessRangeQuery(const meta::hybrid::DataType data_type, const std::string& operand,
                       const query::CompareOperator& com_operator, knowhere::IndexPtr& index_ptr,
                       faiss::ConcurrentBitsetPtr& bitset);
