+++ core/src/scheduler/Definition.h
@@ -35,6 +35,7 @@ using ExecutionEnginePtr = engine::ExecutionEnginePtr;
 using EngineFactory = engine::EngineFactory;
 using EngineType = engine::EngineType;
 using MetricType = engine::MetricType;
+using DataType = engine::meta::hybrid::DataType;
 
 constexpr uint64_t TASK_TABLE_MAX_COUNT = 1ULL << 16ULL;
